fix[#9]: test 코드 폴더 이동
